// src/app/enum/modules.enum.ts
export enum ModulesEnum {
    patientManagement = 1,
    patientCheckout = 2,
    claimsManagement = 3,
    transactions = 4,
    paymentPlans = 5,
    settings = 6,
    insuranceManagement = 7,
    userManagement = 8,
    practitionerManagement = 9,
    formsManagement = 10,
    trainingVideos = 11,
    smsNotifications = 12,
    emailNotifications = 13,
    appointment = 14,
    reports = 15,
    eligibilityManagement = 16,
    roleManagement = 17
}